requirements.txt
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
torch==2.1.0+cu118
torch-lr-finder==0.2.1
torchaudio_augmentations
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
pytest==7.4.3
tensorboard==2.15
tensorboardX==2.6.2.2
matplotlib
pylint
numpy
Pillow
tqdm
requests
pydub